# Optimize NcHelper get primitive array