--- a/WEB-INF/classes/com/cohort/util/StringHolder.java
+++ b/WEB-INF/classes/com/cohort/util/StringHolder.java
@@ -16,6 +16,7 @@ import java.util.Arrays;
 public class StringHolder {
 
     private char[] car; //may be null
+    private final int hashCode;
 
 
     /**
@@ -24,6 +25,7 @@ public class StringHolder {
      */
     public StringHolder(String string) {
         car = string == null? null : string.toCharArray();
+        hashCode = Arrays.hashCode(car); //it allows null and returns 0
     }
 
     /**
@@ -32,6 +34,7 @@ public class StringHolder {
      */
     public StringHolder(char tcar[]) {
         car = tcar;
+        hashCode = Arrays.hashCode(car); //it allows null and returns 0
     }
 
     /**
@@ -60,7 +63,7 @@ public class StringHolder {
      * @return the hashcode for this StringHolder (dependent only on values).
      */
     public int hashCode() {
-        return Arrays.hashCode(car); //it allows null and returns 0
+        return hashCode;
     }
 
     /**