--- a/WEB-INF/classes/gov/noaa/pfel/coastwatch/pointdata/Table.java
+++ b/WEB-INF/classes/gov/noaa/pfel/coastwatch/pointdata/Table.java
@@ -6737,7 +6737,7 @@ Dataset {
                 }
 
                 //get data
-                PrimitiveArray pa = PrimitiveArray.factory(NcHelper.getArray(variable.read())); 
+                PrimitiveArray pa = NcHelper.getPrimitiveArray(variable.read()); 
                 //unpack is done at end of method
                 //String2.log("Table.read4DNc v=" + v + ": " + pa);
 
@@ -6754,7 +6754,7 @@ Dataset {
             //load the stringVariable
             if (stringVariableName != null) {
                 Variable variable = ncFile.findVariable(stringVariableName);
-                PrimitiveArray pa = PrimitiveArray.factory(NcHelper.getArray(variable.read())); 
+                PrimitiveArray pa = NcHelper.getPrimitiveArray(variable.read()); 
                 //unpack is done at end of method
                 if (debugMode) String2.log("  stringVariableName values=" + pa);
                 Attributes atts = new Attributes();
@@ -7034,7 +7034,7 @@ Dataset {
                     tReadShape[nAxes] = variable.getDimension(nAxes).getLength();
                 }
                 Array array = variable.read(tReadOrigin, tReadShape); 
-                PrimitiveArray pa = PrimitiveArray.factory(NcHelper.getArray(array)); 
+                PrimitiveArray pa = NcHelper.getPrimitiveArray(array); 
                 Test.ensureEqual(pa.size(), nRows(),
                     errorInMethod + "Unexpected nRows for " + variable.getFullName() + ".");