EDDTableFromNcFiles.testNThreads();
Pre optimizations:
bigResults:
nThreads=5 time=6269ms
nThreads=4 time=6013ms
nThreads=3 time=6792ms
nThreads=2 time=9124ms
nThreads=1 time=15093ms
nThreads=1 time=16140ms
nThreads=2 time=8768ms
nThreads=3 time=6911ms
nThreads=4 time=6138ms

With optimizations:
bigResults:
nThreads=5 time=4286ms
nThreads=4 time=4258ms
nThreads=3 time=4678ms
nThreads=2 time=5965ms
nThreads=1 time=10441ms
nThreads=1 time=10406ms
nThreads=2 time=5956ms
nThreads=3 time=4723ms
nThreads=4 time=4176ms